requirements.txt
pytest==8.3.3
requests==2.32.3
flake8==7.1.1
pep8-naming==0.14.1
python-dotenv==1.0.1